requirements.txt
better-profanity==0.7.0
discord.py==2.3.2
python-dotenv==1.0.1
